Moves heuristics to own file
