+++ environment/tetris.py
@@ -1,5 +1,6 @@
 import random
 import numpy as np
+from utils.heuristics import Heuristics
 
 line_points = [0, 40, 100, 300, 1200]
 delay_per_level = [48, 43, 38, 33, 28, 23, 18, 13, 8, 6, 5, 5, 5, 4, 4, 4, 3, 3, 3, 2] # <= 19 = 1 
@@ -20,6 +21,7 @@ class Tetris:
         self.rows = rows
         self.high_score = 0
         self.score = 0
+        self.heuristics = Heuristics()
         
         self.reset()
 
@@ -51,43 +53,10 @@ class Tetris:
         self.state = np.array([0, 0, 0, 0])
         return self.state
 
-    # Heuristics
-    def _count_bridges(self, board:np.ndarray):
-        bridge_mask = board != 0
-        empty_count = 0
-
-        for col in range(board.shape[1]):
-            if np.any(bridge_mask[:-1, col]):
-                bridge_row = np.argmax(bridge_mask[:-1, col])
-                empty_count += np.sum(board[bridge_row+1:-1, col] == 0)
-
-        return empty_count
-    
-    def _calculate_height_and_bumpiness(self, board:np.ndarray):
-        heights = np.max(np.where(board != 0, len(board) - 1 - np.arange(len(board))[:, None], 0), axis=0)
-        max_height = np.max(heights)
-        bumpiness = np.sum(np.abs(np.diff(heights)))
-
-        return max_height, bumpiness
-    
-    def _count_full_rows(self, board:np.ndarray):
-        full_rows = np.all(board != 0, axis=1)
-        return int(np.sum(full_rows))
-
     def _get_current_state(self, board:np.ndarray, shape:np.ndarray, offset:tuple):
         new_board = board.copy()
         self._place_shape(new_board, shape, offset)
-        cleared_lines = self._count_full_rows(new_board)
-        holes = self._count_bridges(new_board)
-        bumpiness, height = self._calculate_height_and_bumpiness(new_board)
-
-        return np.array([cleared_lines, holes, bumpiness, height])
-
-    def get_state(self, board:np.ndarray):
-        cleared_lines = self._count_full_rows(board)
-        bridges = self._count_bridges(board)
-        bumpiness, height = self._calculate_height_and_bumpiness(board)
-        return np.array([cleared_lines, bridges, bumpiness, height])
+        return self.heuristics.get_heuristics(new_board)
     
     def get_possible_states(self):
         states = {}
